requirements.txt
Flask==3.1.2
langchain
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
chromadb
requests
beautifulsoup4
pandas
nltk
scikit-learn
tqdm
langchain-groq
python-dotenv
pypdf
pymongo
pandas
langchain-huggingface
sentence-transformers
langchain-chroma
pdfplumber
PyPDF2>=3.0.0
reportlab